event_manager/requirements.txt
asgiref==3.7.2
Django==4.2
django-filter==24.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
phonenumberslite==8.13.31
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
